TOTS-67 Refactored the useGeometryTools to not be a custom hook and not use useCallback/useEffect.
